mtd: cfi_flash: Don't probe if status property is set to "disabled"
Currently, if CONFIG_CFI_FLASH is set, CFI is probed if either "cfi-flash"
or "jedec-flash" property is found in DT, but the property "status" is
not checked.

In some case, it's useful to probe CFI framework depending of "status" DT
property value. For example, a board family which share the same defconfig,
on which a CFI flash is present on some and not on the other. This will
avoid to perform an access to a memory-map area not linked to a CFI device.

This patch checks if "cfi-flash" or "jedec-flash" node's parent has a
status property, if this property is found and set to "disabled", CFI
is not probed.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: Iac8dc68858f2c54d936434c720e6172deb3ef181
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/230304
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/308510
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Christophe KERELLO <christophe.kerello@st.com>