ARM: dts: stm32: Don't probe led-red at boot for stm32mp157a-dk1-u-boot
led-red and button dedicated for fastboot share the same gpio GPIOA13.
Led driver is probed early so the corresponding gpio is taken and
configured in output which forbid fastboot button usage.

To avoid this, remove the "default-state" property which avoid to trigger
the led driver probe() to configure the led default state during startup.

Also remove the status property which is useless.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: Ie7b5a220355bc122ece71eb6382b79386d2ba44c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/364006
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>