ARM: dts: stm32: Don't probe led-red at boot for stm32mp157c-ed1-u-boot
led-red and button dedicated for fastboot share the same gpio GPIOA13.
Led driver is probed early so the corresponding gpio is taken and
configured in output which forbid fastboot button usage.

To avoid this, remove the "default-state" property which avoid to trigger
the led driver probe() to configure the led default state during startup.

Also remove the status property which is useless.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: I8b9b016627025573f73e09c0ce23ff48fb28a824
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/364007
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>