ARM: dts: stm32: Disable WiFi support for stm32mp135f-dk-u-boot
Due to flags CONFIG_EFI_CAPSULE_ON_DISK=y, EFI scan all block device
and try to find an ESP partition on it.
On boards with WiFi support, sdmmc2 is used as interface with WiFi.
When EFI try to find ESP partition on sdmmc2, it sends a mmc request
which fail as WiFi chip doesn't respond.

To avoid this, disable WiFi support in U-Boot.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: I52fc6e747b34ffd2a14419573535e93d706b41c4
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/376241