ARM: dts: stm32: Disable WiFi support for stm32mp157f-dk2-u-boot
Due to flags CONFIG_EFI_CAPSULE_ON_DISK=y, EFI scan all block device
and try to find an ESP partition on it.
On boards with WiFi support, sdmmc2 is used as interface with WiFi.
When EFI try to find ESP partition on sdmmc2, it sends a mmc request
which fail as WiFi chip doesn't respond.

To avoid this, disable WiFi support in U-Boot.

Signed-off-by: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: I2c88718fc1f4bd49bc9e576b5fd35e832c2c44d9
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/376243
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>